Move `Modifiers` to `modifier` package
