+++ src/com/osudroid/ui/v2/game/SliderTicks.kt
@@ -2,7 +2,7 @@ package com.osudroid.ui.v2.game
 
 import com.edlplan.framework.easing.Easing
 import com.reco1l.andengine.Anchor
-import com.reco1l.andengine.Modifiers
+import com.reco1l.andengine.modifier.Modifiers
 import com.reco1l.andengine.container.*
 import com.reco1l.andengine.sprite.*
 import com.reco1l.framework.*
